Fixed issues listed in Issue 51.
Updated build instructions.
Updated various readme files.
Added info how to build the documentation with Doxygen.
Removed all references to encryption and signing.
Removed the extlibs directory.
Removed the maximum version of CMake; this was useful to handle policies automatically.  Now they're done by hand.
Put the minimum version of CMake to 3.15.  Documentation makes references to stuff added in CMake 3.13, latest SWIG & Python find_module() calls were added in CMake 3.12.  Not sure if we make use of that or not.
Added post build steps to api and Python builds to put third party DLLs in place so tests can be run smoothly.
Renamed InterleavedLayer to InterleavedLegacyLayer.
Removed the need to be administrator when building Python.
Made the Python build out of source.
Added example_w_qc_layers.bag; it has ELEVATION and NODE groups.
Travis CI was fixed to use proper environment variables for paths.
