+++ CMakeSettings.json
@@ -7,7 +7,7 @@
             "inheritEnvironments": [ "msvc_x64_x64" ],
             "buildRoot": "${projectDir}\\build_d",
             "installRoot": "${projectDir}\\install_d",
-            "cmakeCommandArgs": "-DCMAKE_EXPORT_COMPILE_COMMANDS:BOOL=ON -DBUILD_PYTHON:BOOL=ON -DCMAKE_PREFIX_PATH=${projectDir}/../third_party/build;${projectDir}/../third_party/build/lib",
+            "cmakeCommandArgs": "-DCMAKE_EXPORT_COMPILE_COMMANDS:BOOL=ON -DBAG_BUILD_PYTHON:BOOL=ON -DCMAKE_PREFIX_PATH=${projectDir}/../third_party/build;${projectDir}/../third_party/build/lib",
             "buildCommandArgs": "-v",
             "ctestCommandArgs": "",
             "intelliSenseMode": "windows-msvc-x64",
@@ -16,6 +16,19 @@
             "clangTidyChecks": "*,-fuchsia-*,-*readability-namespace-comments,-*readability-todo,-google-runtime-references,-*braces-around-statements,-llvm-header-guard,-llvm-namespace-comment,-*readability-braces-around-statements,-*uppercase-literal-suffix,-modernize-use-trailing-return-type",
             "codeAnalysisRuleset": "D:/caris_work/code_analysis_rules.ruleset",
             "variables": []
+        },
+        {
+            "name": "x64-Release",
+            "generator": "Ninja",
+            "configurationType": "Release",
+            "buildRoot": "${projectDir}\\build",
+            "installRoot": "${projectDir}\\install",
+            "cmakeCommandArgs": "-DCMAKE_EXPORT_COMPILE_COMMANDS:BOOL=ON -DBAG_BUILD_PYTHON:BOOL=ON -DCMAKE_PREFIX_PATH=${projectDir}/../third_party/build;${projectDir}/../third_party/build/lib",
+            "buildCommandArgs": "-v",
+            "ctestCommandArgs": "",
+            "inheritEnvironments": [ "msvc_x64_x64" ],
+            "variables": [],
+            "intelliSenseMode": "windows-msvc-x64"
         }
     ]
 }
