+++ api/bag_errors.h
@@ -13,7 +13,6 @@
 
 //! Definitions for error conditions
 #define BAG_GENERAL_ERROR_BASE                    0
-#define BAG_CRYPTO_ERROR_BASE                   200
 #define BAG_METADATA_ERROR_BASE                 400
 #define BAG_HDFV_ERROR_BASE                     600
 #define BAG_COMPOUND_LAYER_ERROR_BASE          1000
@@ -35,14 +34,6 @@ enum BAG_ERRORS {
     BAG_SURFACE_CORRECTIONS_MISSING            =  20, //!< The surface corrections layer is missing
 
     BAG_MORE_BAG_INSTANCES_PRESENT             =  30, //!< There are still instances of the BAG in memory
-
-    BAG_CRYPTO_SIGNATURE_OK                    = 200, //!< Signature found, and valid
-    BAG_CRYPTO_NO_SIGNATURE_FOUND              = 201, //!< No signature found in file
-    BAG_CRYPTO_BAD_SIGNATURE_BLOCK             = 202, //!< Signature found, but invalid
-    BAG_CRYPTO_BAD_KEY                         = 203, //!< Internal key format is invalid
-    BAG_CRYPTO_WRONG_KEY                       = 204, //!< Wrong key type passed
-    BAG_CRYPTO_GENERAL_ERROR                   = 205, //!< Something else went wrong
-    BAG_CRYPTO_INTERNAL_ERROR                  = 206, //!< Something went wrong that the library didn't expect
 
     BAG_METADTA_NO_HOME                        = 400, //!< BAG_HOME directory not set.
     BAG_METADTA_SCHEMA_FILE_MISSING            = 401, //!< Unable to locate schema file.
