+++ api/bag_fordec.h
@@ -10,8 +10,8 @@ class CompoundLayer;
 class CompoundLayerDescriptor;
 class Dataset;
 class Descriptor;
-class InterleavedLayer;
-class InterleavedLayerDescriptor;
+class InterleavedLegacyLayer;
+class InterleavedLegacyLayerDescriptor;
 class Layer;
 class LayerDescriptor;
 class Metadata;
