--- api/readme.txt
@@ -1,13 +1,0 @@
-API
-=======
-
-Don't forget the requirement for BAG_HOME before trying the examples.
-
-On Linux/OS X:
-
-export	BAG_HOME=$PWD/../configdata
-setenv  BAG_HOME $PWD/../configdata
-
-On Windows:
-
-set BAG_HOME=%CD%\..\configdata
