+++ docs/build_instructions.html
@@ -1,2 +1,10 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<html><head><meta content="text/html; charset=ISO-8859-1" http-equiv="content-type"><title>index</title></head><body><h1><img style="width: 1024px; height: 185px;" alt="" src="sub/images/openns_banner.jpg"><br></h1><h1 style="text-align: center;">Build Instructions</h1><a href="sub/visual_studio.html">Visual Studio</a><br><br><a href="sub/mingw.html">MinGW/MSYS</a><br><br><a href="sub/linux.html">Linux</a></body></html>
+<html>
+<head>
+    <meta content="text/html; charset=ISO-8859-1" http-equiv="content-type">
+    <title>index</title>
+</head>
+<body>
+    <h1><img style="width: 1024px; height: 185px;" alt="" src="sub/images/openns_banner.jpg"><br></h1><h1 style="text-align: center;">Build Instructions</h1><a href="sub/visual_studio.html">Visual Studio</a><br><br><a href="sub/linux.html">Linux</a>
+</body>
+</html>
