--- docs/sub/mingw.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<html><head><meta content="text/html; charset=ISO-8859-1" http-equiv="content-type"><title>mingw</title></head><body><h1><a class="mozTocH1" name="mozTocId758748"></a>Building with Mingw/MSYS</h1>Note:
-These instructions assume that the openns source code is located at
-'D:/openns/1.5.0' and the CMake utility is installed at 'C:/cmake'.<br><br><ol id="mozToc"><!--mozToc h2 2 h3 3 h4 4 h5 5 h6 6--><li><a href="#mozTocId297843">Extract the third party modules</a><ol><li><a href="#mozTocId415504">Configure and build zlib</a></li><li><a href="#mozTocId699177">Configure and build HDF5</a></li><li><a href="#mozTocId768644">Configure and build xerces</a></li><li><a href="#mozTocId322545">Configure and build beecrypt</a></li><li><a href="#mozTocId450230">Configure and build bag</a></li></ol></li></ol><h2><a class="mozTocH2" name="mozTocId297843"></a>Extract the third party modules</h2>Browse to the extlibs folder and extract the contained third party
-modules. Note that the version numbers have been removed from the
-folder names after extraction. For the Visual Studio build we will use
-<span style="font-weight: bold;">beecrypt 4.2.1</span>.<br><img style="width: 627px; height: 488px;" alt="" src="images/extract.png"><br><br><h2><a class="mozTocH2" name="mozTocId415504"></a>Configure and build zlib</h2>From the msys command prompt launch the cmake gui to configure the build.<br><img style="width: 573px; height: 251px;" alt="" src="images/launch_cmake_mingw.png"><br><br>Specify the location of the cmake binary files. NOTE - The 'Advanced' option MUST be enabled.<br><img style="width: 647px; height: 612px;" alt="cmake binaries" src="images/zlib_cmake_initial.png"><br><br>Click the 'Configure' button. You will be prompted to create the output folder. Select Yes.<br><img style="width: 338px; height: 149px;" alt="New directory prompt." src="images/zlib_cmake_create_dir.png"><br><br>You will then be prompted to select the appropriate generate.<br><img style="width: 500px; height: 360px;" alt="" src="images/cmake_generator_msys.png"><br><br>Update the following options.<br><br><span style="color: red;">BUILD</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; BUILD_SHARED_LIBS = YES</span><br style="color: red;"><br style="color: red;"><span style="color: red;">CMAKE</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; CMAKE_INSTALL_PREFIX</span><span style="color: red;"> = </span><span style="color: red;">D:/openns/1.5.0/extlibs/zlib/installed</span><br><br>Finally
-click the 'Configure' and then 'Generate' buttons. The makefile is now
-available in the zlib/build folder. From the MSYS console browse to the zlib/build folder and run the following command.<br><br><span style="color: red;">make install</span> <br><br><h2><a class="mozTocH2" name="mozTocId699177"></a>Configure and build HDF5</h2>In the CMake GUI select the HDF5 folder and specify the location of the binary files. Next click the 'Configure' button.<br><img style="width: 615px; height: 538px;" alt="" src="images/hdf5_cmake_initial_vs.png"><br><br>Click the 'Configure' button. You will be prompted to create the output folder. Select Yes.<br><img style="width: 328px; height: 139px;" alt="" src="images/hdf5_cmake_create_dir.png"><br><br>You will then be prompted to select the appropriate generate.<br><img style="width: 500px; height: 360px;" alt="" src="images/cmake_generator_msys.png"><br><br>Update the following options.<br><br style="color: red;"><span style="color: red;">BUILD</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; BUILD_SHARED_LIBS = YES</span><br style="color: red;"><br style="color: red;"><span style="color: red;">CMAKE</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; CMAKE_INSTALL_PREFIX = D:/openns/1.5.0/extlibs/hdf5/installed</span><br><br><span style="color: red;">HDF5</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; HDF5_USE_16_API_DEFAULT = YES</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp;&nbsp;HDF5_ENABLE_Z_LIB_SUPPORT = YES</span><br><br>Click the 'Configure' button again and then update the following options.<br><br><span style="color: red;">ZLIB</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; ZLIB_INCLUDE_DIR = D:/openns/1.5.0/extlibs/zlib/installed/include</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; ZLIB_LIBRARY = D:/openns/1.5.0/extlibs/zlib/installed/lib/zlib.lib</span><br><br>Finally click the 'Configure' and then 'Generate' buttons. The makefile is now
-available in the hdf5/build folder. From the MSYS console browse to the hdf5/build folder and run the following command.<br><br><span style="color: red;">make install</span><br><br><h2><a class="mozTocH2" name="mozTocId768644"></a>Configure and build xerces</h2>From the MSYS console browse to the xerces-c folder and run the follwoing commands:<br><br><span style="color: red;">./configure --prefix=d:/openns/1.5.0/extlibs/xerces-c/installed/</span><br style="color: red;"><span style="color: red;">make -j4 libxerces_c_la_LDFLAGS="-release 3.1 -no-undefined" install</span><br><br><h2><a class="mozTocH2" name="mozTocId322545"></a>Configure and build beecrypt</h2>From the MSYS console browse to the xerces-c folder and run the follwoing commands:<br><br><span style="color: red;">./configure --prefix=d:/openns/1.5.0/extlibs/beecrypt/installed/ --without-java --without-python --disable-static<br>make install</span><br><br><h2><a class="mozTocH2" name="mozTocId450230"></a>Configure and build bag</h2>In the CMake GUI select the openns folder and specify the location of the binary files. Next click the 'Configure' button.<br><img style="width: 616px; height: 539px;" alt="" src="images/openns_cmake_initial_vs.png"><br><br>Click the 'Configure' button. You will be prompted to create the output folder. Select Yes.<br><img style="width: 329px; height: 140px;" alt="" src="images/openns_cmake_create_vs.png"><br><br>You will then be prompted to select the appropriate generate.<br><img style="width: 500px; height: 360px;" alt="" src="images/cmake_generator_msys.png"><br><br>Finally click the 'Configure' and then 'Generate' buttons. The makefile is now
-available in the openns/build folder. From the MSYS console browse to the openns/build folder and run the following command.<br><br><span style="color: red;">make</span></body></html>
