+++ docs/sub/visual_studio.html
@@ -1,2 +1,10 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<html><head><meta content="text/html; charset=ISO-8859-1" http-equiv="content-type"><title>index</title></head><body><h1><img style="width: 1024px; height: 185px;" alt="" src="images/openns_banner.jpg"><br></h1><h1 style="text-align: center;">Build Instructions for Visual Studio</h1><a href="visual_studio_2008.html">Visual Studio 2008</a><br><br><a href="visual_studio_2015.html">Visual Studio 2015</a></body></html>
+<html>
+<head>
+    <meta content="text/html; charset=ISO-8859-1" http-equiv="content-type">
+    <title>index</title>
+</head>
+<body>
+    <h1><img style="width: 1024px; height: 185px;" alt="" src="images/openns_banner.jpg"><br></h1><h1 style="text-align: center;">Build Instructions for Visual Studio</h1><a href="visual_studio_2015.html">Visual Studio 2015</a><br><br><a href="visual_studio_2017.html">Visual Studio 2017</a>
+</body>
+</html>
