--- docs/sub/visual_studio_2008.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<html><head><meta content="text/html; charset=ISO-8859-1" http-equiv="content-type"><title>visual_studio</title></head><body><h1><a class="mozTocH1" name="mozTocId758748"></a>Building with Microsoft Visual Studio 2008</h1>Note:
-These instructions assume that the openns source code is located at
-'D:/openns/1.5.0' and the CMake utility is installed at 'C:/cmake'.<br><br><ol id="mozToc"><li><a href="#mozTocId297843">Extract the third party modules</a><ol><li><a href="#mozTocId543006">Configure and build zlib</a></li><li><a href="#mozTocId699177">Configure and build HDF5</a></li><li><a href="#mozTocId768644">Configure and build libxml2</a></li><li><a href="#mozTocId322545">Configure and build beecrypt</a></li><li><a href="#mozTocId645668">Configure and build bag<span style="font-weight: bold;"></span></a></li></ol></li></ol><h2><a class="mozTocH2" name="mozTocId297843"></a>Extract the third party modules</h2>Browse
-to the extlibs folder and extract the contained third party modules.
-Note that the version numbers have been removed from the folder names
-after extraction. For the Visual Studio build we will use <span style="font-weight: bold;">beecrypt 4.1.2</span>.<br><img style="width: 627px; height: 488px;" alt="" src="images/extract.png"><br><br><br><h2><a class="mozTocH2" name="mozTocId543006"></a>Configure and build zlib</h2>From the Visual Studio command prompt launch the cmake gui to configure the build.<br><img style="width: 565px; height: 215px;" alt="Launch cmake" src="images/launch_cmake_vs.png"><br><br>Specify the location of the cmake binary files. NOTE - The 'Advanced' option MUST be enabled.<br><img style="width: 647px; height: 612px;" alt="cmake binaries" src="images/zlib_cmake_initial.png"><br><br>Click the 'Configure' button. You will be prompted to create the output folder. Select Yes.<br><img style="width: 338px; height: 149px;" alt="New directory prompt." src="images/zlib_cmake_create_dir.png"><br><br>You will then be prompted to select the appropriate generate. NOTE - Be sure to select the Win64 generator for x64 builds.<br><img style="width: 500px; height: 360px;" alt="" src="images/cmake_generator_vs.png"><br><br>Select the 'BUILD_SHARED_LIBS' option and specify the installation prefix.<br><br><span style="color: red;">BUILD</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; BUILD_SHARED_LIBS = YES</span><br style="color: red;"><br style="color: red;"><span style="color: red;">CMAKE</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; CMAKE_INSTALL_PREFIX = D:/openns/1.5.0/extlibs/zlib/installed</span><br><br>Finally
-click the 'Configure' and then 'Generate' buttons. The solution is now
-available in the zlib/build folder. Launch Visual Studio and open the
-newly generated solution. Build the solution in Release and Debug. When
-the build has completed build the 'Install' project in Release and Debug. <br><br><br><h2><a class="mozTocH2" name="mozTocId699177"></a>Configure and build HDF5</h2>In the CMake GUI select the HDF5 folder and specify the location of the binary files. Next click the 'Configure' button.<br><img style="width: 615px; height: 538px;" alt="" src="images/hdf5_cmake_initial_vs.png"><br><br>Click the 'Configure' button. You will be prompted to create the output folder. Select Yes.<br><img style="width: 328px; height: 139px;" alt="" src="images/hdf5_cmake_create_dir.png"><br><br>You will then be prompted to select the appropriate generate. NOTE - Be sure to select the Win64 generator for x64 builds.<br><img style="width: 500px; height: 360px;" alt="" src="images/cmake_generator_vs.png"><br><br>Update the following options.<br><br style="color: red;"><span style="color: red;">BUILD</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; BUILD_SHARED_LIBS =&nbsp;YES</span><br style="color: red;"><br style="color: red;"><span style="color: red;">CMAKE</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; CMAKE_INSTALL_PREFIX = D:/openns/1.5.0/extlibs/hdf5/installed</span><br><br><span style="color: red;">HDF5</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; HDF5_USE_16_API_DEFAULT = YES</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp;&nbsp;HDF5_ENABLE_Z_LIB_SUPPORT = YES</span><br><br>Click the 'Configure' button again and then update the following options.<br><br><span style="color: red;">ZLIB</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; ZLIB_INCLUDE_DIR = D:/openns/1.5.0/extlibs/zlib/installed/include</span><br style="color: red;"><span style="color: red;">&nbsp;&nbsp;&nbsp; ZLIB_LIBRARY = optimized;D:/openns/1.5.0/extlibs/zlib/installed/lib/zlib.lib;debug;D:/openns/1.5.0/extlibs/zlib/installed/lib/zlib_d.lib</span><br><br>Finally click the 'Configure' and then 'Generate' buttons. The solution
-is now available in the hdf5/build folder. Launch Visual Studio and
-open the newly generated solution. Build the solution in Release and Debug. When the build has completed build the 'Install' project in
-Release and Debug.<br><br><h2><a class="mozTocH2" name="mozTocId768644"></a>Configure and build libxml2</h2>
-Launch the Visual Studio command prompt and browse to the libxml2/win32 folder and run the following command:<br><br><span style="color: red;">cscript configure.js compiler=msvc debug=no vcmanifest=yes iconv=no sodir=D:\openns\1.5.0\extlibs\libxml2\bin\release libdir=D:\openns\1.5.0\extlibs\libxml2\lib\release</span><br><span style="color: red;">nmake /f Makefile.msvc clean</span><br><span style="color: red;">nmake /f Makefile.msvc</span><br><span style="color: red;">nmake /f Makefile.msvc install</span>
-<br><br><span style="color: red;">cscript configure.js compiler=msvc debug=yes vcmanifest=yes cruntime=/MDd iconv=no sodir=D:\openns\1.5.0\extlibs\libxml2\bin\debug libdir=D:\openns\1.5.0\extlibs\libxml2\lib\debug</span><br><span style="color: red;">nmake /f Makefile.msvc clean</span><br><span style="color: red;">nmake /f Makefile.msvc</span><br><span style="color: red;">nmake /f Makefile.msvc install</span>
-<h2><a class="mozTocH2" name="mozTocId322545"></a><br>Configure and build beecrypt</h2>Launch the Visual Studio command prompt and browse to the beecrypt folder and run the following command:<br><br><span style="color: red;">nmake -f Makefile_bag X64=TRUE DEBUG=FALSE</span><br><span style="color: red;">nmake -f Makefile_bag clean</span><br><span style="color: red;">nmake -f Makefile_bag X64=TRUE DEBUG=TRUE</span><h2><a class="mozTocH2" name="mozTocId645668"></a><br>Configure and build bag</h2>In the CMake GUI select the openns folder and specify the location of the binary files. Next click the 'Configure' button.<br><img style="width: 616px; height: 539px;" alt="" src="images/openns_cmake_initial_vs.png"><br><br>Click the 'Configure' button. You will be prompted to create the output folder. Select Yes.<br><img style="width: 329px; height: 140px;" alt="" src="images/openns_cmake_create_vs.png"><br><br>You will then be prompted to select the appropriate generate. NOTE - Be sure to select the Win64 generator for x64 builds.<br><img style="width: 500px; height: 360px;" alt="" src="images/cmake_generator_vs.png"><br><br>Finally click the 'Configure' and then 'Generate' buttons. The solution
-is now available in the openns/build folder. Launch Visual Studio and
-open the newly generated solution. Build the solution in Release and Debug.</body></html>
