+++ docs/sub/visual_studio_2017.html
@@ -1,0 +1,72 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
+<html>
+<head>
+	<meta content="text/html; charset=ISO-8859-1" http-equiv="content-type">
+	<title>linux</title>
+</head>
+<body>
+	<h1>Building with Visual Studio 2017</h1>
+	Note:  It is assumed you install your third party libraries into C:\third_party.<br>
+	Note:  Debug versions can be built as well.  Replace -DCMAKE_BUILD_TYPE=Release with -DCMAKE_BUILD_TYPE=Debug.  For LibXML2, set debug=yes.<br>
+	Note:  Before configuring with CMake, make sure to setup the Visual Studio developer environment by running C:\Program Files (x86)\Microsoft Visual Studio\2017\EDITION\VC\Auxiliary\Build\vcvars64.bat.<br>
+
+	<h2>Configure, build and install ZLib</h2>
+	Get the newest 1.2.x version from https://zlib.net/<br>
+	In the directory you unzipped the source package you downloaded, type the following:<br>
+	<br>
+	<span style="color: red;">
+		cmake -B build -G "Visual Studio 15 2017 Win64" -S . -DCMAKE_INSTALL_PREFIX:PATH=C:\third_party -DCMAKE_BUILD_TYPE=Release<br>
+		cmake --build build --target install<br>
+	</span>
+
+	<h2>Configure, build and install LibXML2</h2>
+	Get the newest 2.9.x version from ftp://xmlsoft.org/libxml2/<br>
+	In the directory you unzipped the source package you downloaded, type the following:<br>
+	<br>
+	<span style="color: red;">
+		set opts=zlib=yes iconv=no prefix=C:\third_party sodir=C:\third_party\bin include=C:\third_party\include lib=C:\third_party\lib<br>
+		<br>
+		cscript configure.js %opts% debug=no cruntime=/MD<br>
+		nmake -nologo -f Makefile clean<br>
+		nmake -nologo -f Makefile install-libs<br>
+	</span>
+
+	<h2>Configure, build and install HDF5</h2>
+	Get the newest 1.12.x version from https://www.hdfgroup.org/downloads/hdf5/source-code/<br>
+	In the directory you unzipped the source package you downloaded, type the following:<br>
+	<br>
+	<span style="color: red;">
+		cmake -B build -G "Visual Studio 15 2017 Win64" -S . -DCMAKE_INSTALL_PREFIX:PATH=C:\third_party -DCMAKE_BUILD_TYPE=Release -DHDF5_BUILD_TOOLS:BOOL=OFF -DBUILD_TESTING:BOOL=OFF -DBUILD_SHARED_LIBS:BOOL=ON -DHDF5_BUILD_HL_LIB:BOOL=ON -DHDF5_ENABLE_Z_LIB_SUPPORT:BOOL=ON -DZLIB_INCLUDE_DIRS=C:\third_party\include -DZLIB_LIBRARY:PATH="optimized;C:\third_party\lib\zlib.lib"<br>
+		<br>
+		cmake --build build --target install<br>
+	</span>
+
+	<h2>Configure, build and install Catch2</h2>
+	Get the newest 2.x version from https://github.com/catchorg/Catch2/releases<br>
+	In the directory you unzipped the source package you downloaded, type the following:<br>
+	<br>
+
+	<span style="color: red;">
+		cmake -B build -G "Visual Studio 15 2017 Win64" -S . -DCMAKE_INSTALL_PREFIX:PATH=C:\third_party -DBUILD_TESTING:BOOL=OFF<br>
+		<br>
+		cmake --build build --target install<br>
+	</span>
+
+	<h2>Install SWIG (optional; required for Python support)</h2>
+	Get the swigwin package from http://www.swig.org/download.html<br>In the directory you unzipped the package you downloaded, type the following:<br>
+	<br>
+	<span style="color: red;">
+		copy swig.exe C:\third_party\bin<br>
+		xcopy /s Lib C:\third_party\lib<br>
+	</span>
+
+	<h2>Configure, build and install BAG</h2>
+	To generate the .vcxproj and .sln files, simply go to the directory you unzipped the archive into, and type:<br>
+	<br>
+	<span style="color: red;">cmake -G "Visual Studio 15 2017" -B build -S . -DCMAKE_BUILD_CONFIG=Release -DCMAKE_PREFIX_PATH=C:\third_party -DCMAKE_INSTALL_PREFIX=C:\third_party</span><br>
+	<br>
+	If Python support is needed, SWIG 4.0.1 or newer is required.  Simply add <span style="color: red;">-DBUILD_PYTHON=ON</span> to the above command.
+	<br>
+	The solution file is now available in the build directory. Open it with Visual Studio 2017, select the configuration (Debug or Release), and start a build.:<br>
+</body>
+</html>
