--- extlibs/readme
@@ -1,8 +1,0 @@
-   EXTLIBS
-   =======
-   
-This directory contains a number of external libraries which are used
-by the Open Navigation Surface project. While you can download and
-build them all separately they are included here in order to make it
-easier to get a quick start. It is possible to substitute other versions
-or completely different libraries as long as they maintain the same calling interface.
